Sleep for, or until a given time

This package provides an internal clock measured in milliseconds since the last reset, with functions to get the current timestamp, reset it to zero, sleep for a number of milliseconds, or sleep until the relative time is beyond a certain number of milliseconds. All functions return the relative time in millseconds at the end of the function call.

This aim is to provide a lower overhead alternative to Sys.sleep, for the purpose of simple animation at a reasonably consistent frame-rate, regardless of platform and system.

wait_for(delay_ms)

wait_until(time_ms)

waitr_timestamp()

waitr_reset()

Arguments

delay_ms
Milliseconds to wait.

time_ms
Clock time to wait for, in milliseconds.

Value

The time in milliseconds, relative to the last reset.